+++ pkg/statistics/BUILD.bazel
@@ -74,7 +74,7 @@ go_test(
     data = glob(["testdata/**"]),
     embed = [":statistics"],
     flaky = True,
-    shard_count = 34,
+    shard_count = 32,
     deps = [
         "//pkg/config",
         "//pkg/parser/ast",
